% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{BackwardPass.HiddenMarkovModel}
\alias{BackwardPass.HiddenMarkovModel}
\title{Backward pass in a hidden Markov model}
\usage{
\method{BackwardPass}{HiddenMarkovModel}(
  H = parent.frame()$H,
  idx = parent.frame()$idx - 1,
  verbose = parent.frame()$verbose
)
}
\description{
This is an internal function of the package \code{hidden}.
It implements the recursive backward algorithm for computing posterior probability distributions of hidden states in hidden Markov models.
It is called by the method \code{Smooth.HiddenMarkovModel}.
}
