% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{Filter.HiddenMarkovModel}
\alias{Filter.HiddenMarkovModel}
\title{Apply filtering to a HiddenMarkovModel object}
\usage{
\method{Filter}{HiddenMarkovModel}(H, idx, verbose = TRUE)
}
\arguments{
\item{H}{\code{HiddenMarkovModel} object, as created by the constructor function \code{HiddenMarkovModel}.}

\item{idx}{index of the last time step for which to run filtration (filtration probabilities for the ones preceding it will be stored automatically).}

\item{verbose}{logical; whether to report progression of the forward pass.}
}
\description{
Applies the filtering algorithm by a recursive forward pass through a hidden Markov model.
This lets us compute the posterior probability distribution of the most recent state given all evidence up to a current point in time.
A vector of observations needs to be included with the model for this algorithm to work (use \code{SetObservations.HiddenMarkovModel}).
It is acceptable for the vector of observations to include \code{NA}s at the end (then, we apply prediction instead of filtration at some point).
}
\details{
After the algorithm finishes, see \code{YourHiddenMarkovModelObject$probs[[YourTimePointIndex]]$filtered} to retrieve the resulting probability distribution.
}
