% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{ForwardPass.HiddenMarkovModel}
\alias{ForwardPass.HiddenMarkovModel}
\title{Forward pass in a hidden Markov model}
\usage{
\method{ForwardPass}{HiddenMarkovModel}(
  H = parent.frame()$H,
  idx = parent.frame()$idx - 1,
  verbose = parent.frame()$verbose
)
}
\description{
This is an internal function of the package \code{hidden}.
It implements the recursive forward algorithm for computing posterior probability distributions of hidden states in hidden Markov models.
It is called by the methods \code{Filter.HiddenMarkovModel} and \code{Smooth.HiddenMarkovModel}.
}
