% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.R
\name{H}
\alias{H}
\title{Hide non-function variables from function.}
\usage{
H(...)
}
\arguments{
\item{...}{Named functions and function definitions.}
}
\value{
Nothing.
}
\description{
Hide non-function variables from function.
}
