% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{HiddenMarkovModel}
\alias{HiddenMarkovModel}
\title{Construct a HiddenMarkovModel object}
\usage{
HiddenMarkovModel(
  Tr,
  Em,
  X0,
  hidden_states = NULL,
  observables = NULL,
  name = paste0("HMM_", as.character(Sys.Date()))
)
}
\arguments{
\item{Tr}{transition matrix; a \code{K}-by-\code{K} numeric matrix where \code{K} is the number of possible hidden states.}

\item{Em}{emission matrix; a \code{K}-by-\code{N} numeric matrix where \code{N} is the number of possible observable states.}

\item{X0}{initial probability distribution; a column vector of initial probabilities of hidden states at time zero.}

\item{hidden_states;}{character vector of names of hidden states in order. If not specified, they are named '\code{x1}', '\code{x2}', ... , '\code{xK}'.}

\item{observables;}{character vector of names of possible observable states in order. If not specified, they are named '\code{e1}', '\code{e2}', ... , '\code{eN}'.}

\item{name;}{internal name of the \code{HiddenMarkovModel} object. If not specified, it is named '\code{HMM_}' followed by the system date.}
}
\description{
Constructs an object for simulation of a hidden Markov model.
This can be used to apply the filtration, prediction, smoothing and Viterbi algorithms.
In the future, a plotting capability might also be added.
}
