% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{plot.HiddenMarkovModel}
\alias{plot.HiddenMarkovModel}
\title{Plot a HiddenMarkovModel object diagram}
\usage{
\method{plot}{HiddenMarkovModel}(H, ...)
}
\description{
Plot the transition and observation model of a \code{HiddenMarkovModel} object using package \code{igraph}.
}
