% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{print.HiddenMarkovModel}
\alias{print.HiddenMarkovModel}
\title{Print information about a HiddenMarkovModel object}
\usage{
\method{print}{HiddenMarkovModel}(H)
}
\description{
Prints information about the properties of a \code{HiddenMarkovModel} object.
}
