% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{PrintStats.HiddenMarkovModel}
\alias{PrintStats.HiddenMarkovModel}
\title{Report HiddenMarkovModel object statistics}
\usage{
\method{PrintStats}{HiddenMarkovModel}(H)
}
\description{
Prints computed statistics stored in a \code{HiddenMarkovModel} object: probabilities obtained by forward pass, backward pass, smoothing and the Viterbi algorithm.
}
