% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{ResetStats.HiddenMarkovModel}
\alias{ResetStats.HiddenMarkovModel}
\title{Delete computed statistics in a HiddenMarkovModel object}
\usage{
\method{ResetStats}{HiddenMarkovModel}(H)
}
\arguments{
\item{H}{\code{HiddenMarkovModel} object, as created by the constructor function \code{HiddenMarkovModel}}
}
\description{
Clears fields \code{probs}, \code{most_likely_paths_probs}, \code{most_likely_paths_pointers} and \code{most_likely_path}.
}
