% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{SetObservations}
\alias{SetObservations}
\title{Add observations to a HiddenMarkovModel object}
\usage{
SetObservations(H, values)
}
\arguments{
\item{H}{\code{HiddenMarkovModel} object, as created by the constructor function \code{HiddenMarkovModel}.}

\item{values}{character or integer vector; either a vector of observable state names (taken from the names of observables specified for \code{H}) or a vector of their indices with respect to the order in which they are listed (use \code{print(H)} to check).}
}
\description{
Adds a vector of observations at progressing time points to a \code{HiddenMarkovModel} object.
Please beware: this erases any stats you have computed on your model so far.
}
