% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{Smooth.HiddenMarkovModel}
\alias{Smooth.HiddenMarkovModel}
\title{Apply smoothing to a HiddenMarkovModel object}
\usage{
\method{Smooth}{HiddenMarkovModel}(H, idcs, verbose = TRUE)
}
\arguments{
\item{H}{\code{HiddenMarkovModel} object, as created by the constructor function \code{HiddenMarkovModel}.}

\item{idcs}{a single index or a vector of indices of time points for which smoothing is to be done.}

\item{verbose}{logical; whether to report progression of the forward pass and the backward pass.}
}
\description{
Applies the smoothing algorithm by a recursive forward pass and backward passthrough a hidden Markov model.
This lets us compute the posterior probability distribution of a past state given all evidence up to a current point in time.
A vector of observations needs to be included with the model for this algorithm to work (use \code{SetObservations.HiddenMarkovModel}).
It is not acceptable for the vector of observations to include any \code{NA}s.
}
\details{
After the algorithm finishes, see \code{YourHiddenMarkovModelObject$probs[[YourTimePointIndex]]$smoothed} to retrieve the resulting probability distribution.
If interested in the probability distributions computed in the forward pass, see the slot \code{filtered} instead.
To see the probability distributions computed in the backward pass, see the slot \code{backward}.
}
