% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{Viterbi.HiddenMarkovModel}
\alias{Viterbi.HiddenMarkovModel}
\title{Apply the Viterbi algorithm to a HiddenMarkovModel object}
\usage{
\method{Viterbi}{HiddenMarkovModel}(H, idx.end = H$n_observations)
}
\arguments{
\item{H}{\code{HiddenMarkovModel} object, as created by the constructor function \code{HiddenMarkovModel}.}

\item{idx.end}{an index of the time points where path terminates.}

\item{verbose}{logical; whether to report progression of the Viterbi algorithm.}
}
\description{
Applies the Viterbi algorithm using dynamic programming on a hidden Markov model.
This lets us compute the most likely progression (path) of hidden states given a sequence of observable states.
A vector of observations needs to be included with the model for this algorithm to work (use \code{SetObservations.HiddenMarkovModel}).
It is not acceptable for the vector of observations to include any \code{NA}s.
}
\details{
After the algorithm finishes, see \code{YourHiddenMarkovModelObject$most_likely_path} to retrieve the most likely sequence of hidden states.
If interested in the dynamic programming matrix with path probabilities, see the slot \code{most_likely_paths_probs} instead.
To view the matrix with pointers for backtracking in the Viterbi algorithm, see the slot \code{most_likely_path_pointers}.
}
